Move @kksh/jsr package to @kksh/api
