+++ packages/api/jsr.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://jsr.io/schema/config-file.v1.json",
 	"name": "@kunkun/api",
-	"version": "0.0.48",
+	"version": "0.0.49",
 	"license": "MIT",
 	"exports": {
 		".": "./src/index.ts",
@@ -15,7 +15,8 @@
 		"./supabase": "./src/supabase/index.ts",
 		"./supabase/types": "./src/supabase/database.types.ts",
 		"./dev": "./src/dev/index.ts",
-		"./events": "./src/events.ts"
+		"./events": "./src/events.ts",
+		"./extensions/jsr": "./src/extensions/jsr/index.ts"
 	},
 	"imports": {}
 }
