+++ packages/api/package.json
@@ -16,7 +16,8 @@
 		"./events": "./src/events.ts",
 		"./supabase": "./src/supabase/index.ts",
 		"./supabase/types": "./src/supabase/database.types.ts",
-		"./package.json": "./package.json"
+		"./package.json": "./package.json",
+		"./extensions/jsr": "./src/extensions/jsr/index.ts"
 	},
 	"license": "MIT",
 	"scripts": {
@@ -41,6 +42,7 @@
 		"typescript": "^5.0.0"
 	},
 	"dependencies": {
+		"@hk/jsr-client": "npm:@jsr/hk__jsr-client@^0.1.2",
 		"@tauri-apps/api": "^2.2.0",
 		"@tauri-apps/cli": "^2.2.2",
 		"@tauri-apps/plugin-deep-link": "^2.2.0",
