--- packages/jsr/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "@kksh/jsr",
-	"type": "module",
-	"scripts": {
-		"test": "bun test --coverage"
-	},
-	"exports": {
-		".": "./src/index.ts",
-		"./models": "./src/models.ts"
-	},
-	"devDependencies": {
-		"@types/bun": "latest"
-	},
-	"peerDependencies": {
-		"@kksh/api": "workspace:*",
-		"typescript": "^5.0.0"
-	},
-	"dependencies": {
-		"@hk/jsr-client": "npm:@jsr/hk__jsr-client@^0.1.2",
-		"valibot": "1.0.0-beta.11"
-	}
-}
